Extract standardEvent action definition into a reusable factory
- Add action.ts exporting standardEventAction(trackingType, title,
  description, defaultSubscription?) - builds the shared fields/
  perform/performBatch logic for the standardEvent action family.
  When trackingType is set, tracking_type is hardcoded and hidden
  from the UI (for future preset-replacement actions); when
  undefined, tracking_type stays user-selectable (current
  standardEvent behavior)
- standardEvent/index.ts is now a one-line call into the factory
- Move isCanary/resolveVersion from v3/utils-v3.ts into action.ts
  (FLAGON_NAME stays in v3/utils-v3.ts); customEvent/index.ts now
  imports resolveVersion from ../action instead
