+++ packages/destination-actions/src/destinations/reddit-conversions-api/v3/utils-v3.ts
@@ -1,37 +1,17 @@
-import type { Features, RequestClient } from '@segment/actions-core'
+import type { RequestClient } from '@segment/actions-core'
 import { PayloadValidationError } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload as StandardEvent } from '../standardEvent/generated-types'
 import type { Payload as CustomEvent } from '../customEvent/generated-types'
 import { V3EventItem, V3Payload, V3Metadata, V3Product, V3User, V3DataProcessingOptions } from './types-v3'
 import { processHashing } from '../../../lib/hashing-utils'
-import { LEGACY_API_VERSION, LATEST_API_VERSION, ApiVersion } from '../versioning-info'
+import { LATEST_API_VERSION } from '../versioning-info'
 
 /** FLAGON_NAME
  * Flagon flag gating whether Reddit Conversions API v3 is reachable at all for an account.
  * https://flagon.segment.com/families/centrifuge-destinations/gates/reddit-conversions-api-canary-version
  */
 export const FLAGON_NAME = 'reddit-conversions-api-canary-version'
-
-export function isCanary(features?: Features): boolean {
-  return !!(features && features[FLAGON_NAME])
-}
-
-/** resolveVersion
- * Decides which Reddit Conversions API version a given event should use.
- *
- * The Flagon flag is the ops-level rollout/kill-switch: if it's off, every event stays on v2,
- * regardless of the customer's `api_version` field selection. If the flag is on, the customer's
- * per-action `api_version` field decides: existing customers who configured this action before
- * the field existed have no value for it (resolves to `undefined` here), so they stay on v2 too,
- * until they explicitly opt into `'v3'`.
- */
-export function resolveVersion(apiVersion: string | undefined, features?: Features): ApiVersion {
-  if (!isCanary(features)) {
-    return LEGACY_API_VERSION
-  }
-  return apiVersion === LATEST_API_VERSION ? LATEST_API_VERSION : LEGACY_API_VERSION
-}
 
 type EventMetadataType = StandardEvent['event_metadata'] | CustomEvent['event_metadata']
 type ProductsType = StandardEvent['products'] | CustomEvent['products']
